feat: ♻️ Prepares widgets and panel for container layout.
Refactors chunks of ui code by removing UIElement class, therefore a Container is now a standalone class and sizing depends only on the "parent" PanelBase instance. Add methods to panel to force pack operation to keep widgets position according to the layout. Added methods for removing children based on their ID and to add children to the Screen.